FOR_UPSTREAM [VPG]: drm/i915/gen8: Fix sporadic GFX page fault issue
After dynamic page allocation was enabled, page faults were observed in the
OGL conformance test suite. After debugging, it was found that there was an out
of bound access, caused by the 3D driver.

This out of bound access no longer happened after resource streamer was enabled,
but in the case CHV prefetches unmapped pages, it will trigger unrecoverable
faults. Therefore, we must ensure the entire address space is mapped.

This wasn't a problem before because in aliased and full PPGTT (before dynamic
alloc) mode, we initialize all the PTEs with a scratch page. But with deferred
allocation, we program only the used PTEs, so the unused entries remain
uninitialized.

As a failsafe solution, will allocate a scratch Page table also for each context,
make all the entries of scratch PD point to it. And then make all entries of scratch
page table point to the scratch page.

v2: Do not leak scratch_pt.

Change-Id: I74ccdcd497dc0676b1073faf5f2c1d5fdeb0b0e3
Issue: GMINL-5449
Issue: VIZ-4619
Signed-off-by: Akash Goel <akash.goel@intel.com>
Signed-off-by: Michel Thierry <michel.thierry@intel.com>