iio: accel: bmc150: use available_scan_masks
Use available_scan_masks to allow the iio core to select
the data to send to userspace depending on which axes are
enabled, instead of doing this in the driver's interrupt
handler.

This also fixes the issue of accessing the buffer scan_mask
instead of active_scan_mask, since these might not be the
same due to client devices.

Change-Id: Id4e58d4426bc38888504ceb721f4de5e75590c8c
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4428
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>